rename packages and delete generate antlr files
